refactor: create a service to validate inputs in InputParser
