+++ src/main/java/com/om1cael/Main.java
@@ -1,10 +1,6 @@
 package com.om1cael;
 
-import com.om1cael.controller.MovieController;
 import com.om1cael.view.MenuView;
-import com.om1cael.view.InputParser;
-
-import java.util.Scanner;
 
 public class Main {
     public static void main(String[] args) {
