+++ src/main/java/com/om1cael/controller/MovieController.java
@@ -3,7 +3,6 @@ package com.om1cael.controller;
 import com.om1cael.model.Movie;
 import com.om1cael.services.MovieService;
 import com.om1cael.view.FeedbackView;
-import com.om1cael.view.InputParser;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -27,12 +26,13 @@ public class MovieController {
         System.out.println("Adding a new movie!");
         String movieTitle = this.inputParser.getTextInput(" - Title: ");
         String movieGenre = this.inputParser.getTextInput(" - Genre: ");
-        int releaseYear = this.inputParser.getNumberInput(" - Release year: ", 1800, LocalDate.now().getYear());
-        byte rating = (byte)this.inputParser.getNumberInput(" - Rating (0-5): ", 0, 5);
+        int releaseYear = this.inputParser.getNumberInput(" - Release year: ", 1800, LocalDate.now().getYear(), false);
+        byte rating = (byte)this.inputParser.getNumberInput(" - Rating (0-5): ", 0, 5, false);
         int movieStatusChoice = this.inputParser.getNumberInput(
                 " - Status (1 - Watched, 2 - To watch, 3 - Not interested in): ",
                 1,
-                3
+                3,
+                false
         );
 
         Movie movie = this.movieService.addMovie(
@@ -58,14 +58,14 @@ public class MovieController {
     }
 
     public void filterByYear() {
-        int year = this.inputParser.getNumberInput("Year: ", 1800, LocalDate.now().getYear());
+        int year = this.inputParser.getNumberInput("Year: ", 1800, LocalDate.now().getYear(), false);
         List<Movie> filteredList = this.movieService.filterByYear(year);
         feedbackView.printFilterFeedback(filteredList,
                 "List of movies released in " + year);
     }
 
     public void filterByRating() {
-        byte rating = (byte) this.inputParser.getNumberInput("Rating: ", 0, 5);
+        byte rating = (byte) this.inputParser.getNumberInput("Rating: ", 0, 5, false);
         List<Movie> filteredList = this.movieService.filterByRating(rating);
         feedbackView.printFilterFeedback(filteredList,
                 "List of movies with rating of " + rating);
